+++ tests-e2e/publish/3-upload.spec.ts
