+++ tests-e2e/publish/meta-collection.spec.ts
@@ -2,7 +2,7 @@ import { test, expect } from '@playwright/test';
 
 import { publisherContext } from '../../playwright/.auth/contexts';
 import { appConfig } from '../../src/config';
-import { metadataA as dataset } from '../fixtures/datasets';
+import { metadataA as dataset, metadataB as datasetB } from '../fixtures/datasets';
 
 import { CollectionPage } from './pages/collection-page';
 
@@ -26,21 +26,23 @@ test.describe('Metadata Data Collection', () => {
   test.describe('Authed as a publisher', () => {
     test.use({ storageState: publisherContext });
 
-    test.beforeEach(async () => {
-      await collectionPage.goto(dataset.id);
-    });
-
     test('Has a heading', async ({ page }) => {
+      await collectionPage.goto(dataset.id);
       await expect(page.getByRole('heading', { name: 'How was the data collected or calculated?' })).toBeVisible();
     });
 
     test.fixme('Can switch to Welsh', async ({ page }) => {
       // TODO: waiting on translations
+      await collectionPage.goto(dataset.id);
       await page.getByText('Cymraeg').click();
       await expect(page.getByRole('heading', { name: '' })).toBeVisible();
     });
 
-    test.describe('Form', () => {
+    test.describe('Form validation', () => {
+      test.beforeEach(async () => {
+        await collectionPage.goto(dataset.id);
+      });
+
       test('Displays a validation error when no input is provided', async ({ page }) => {
         await collectionPage.fillForm('');
         await collectionPage.submit();
@@ -54,11 +56,17 @@ test.describe('Metadata Data Collection', () => {
         await expect(page.url()).toBe(`${baseUrl}/en-GB/publish/${dataset.id}/collection`);
         await expect(page.getByText('Enter how the data was collected or calculated')).toBeVisible();
       });
+    });
+
+    test.describe('Form validation', () => {
+      test.beforeEach(async () => {
+        await collectionPage.goto(datasetB.id);
+      });
 
       test('Can add info about data collection and return to the tasklist', async ({ page }) => {
         await collectionPage.fillForm('This is info about collection.');
         await collectionPage.submit();
-        await expect(page.url()).toBe(`${baseUrl}/en-GB/publish/${dataset.id}/tasklist`);
+        await expect(page.url()).toBe(`${baseUrl}/en-GB/publish/${datasetB.id}/tasklist`);
       });
     });
   });
