+++ tests-e2e/publish/pages/collection-page.ts
