+++ tests-e2e/publish/pages/provider-page.ts
