+++ tests-e2e/publish/pages/related-page.ts
