--- tests/sample-csvs/missing-column-header.csv
@@ -1,3 +1,0 @@
-id,text,,geo
-1,test 1,4532,xy
-2,test 2,4348,zf
