--- tests/sample-csvs/test-data-1.csv
@@ -1,3 +1,0 @@
-id,text,number
-1,test 1,4532
-2,test 2,4348
